refactoring

+++ gpu/vadd.cl
@@ -1,4 +1,4 @@
-__kernel void vector_add(__global const int *A, __global const int *B, __global int *C) 
+__kernel void vector_add(__global const float *A, __global const float *B, __global float *C) 
 {
   int i = get_global_id(0);
   C[i] = A[i] + B[i];
